Organized HTML files and controllers -- did not touch css or js files
