+++ src/main/java/org/wecancodeit/sketchflex/exceptions/SketchDeckNotFoundException.java
@@ -1,0 +1,14 @@
+package org.wecancodeit.sketchflex.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+public class SketchDeckNotFoundException extends Exception {
+
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 7731571911415429108L;
+
+}
