+++ src/main/java/org/wecancodeit/sketchflex/storage/StorageService.java
@@ -5,6 +5,8 @@ import java.util.stream.Stream;
 
 import org.springframework.core.io.Resource;
 import org.springframework.web.multipart.MultipartFile;
+import org.wecancodeit.sketchflex.exceptions.StorageException;
+import org.wecancodeit.sketchflex.exceptions.StorageFileNotFoundException;
 
 public interface StorageService {
 	void init() throws StorageException;
