+++ src/main/resources/static/css/style.css
@@ -369,7 +369,3 @@ h1{
    }
 
 }
-
-
-
-
