+++ src/main/resources/templates/index.html
@@ -1,51 +1,49 @@
 <!DOCTYPE html>
 <html lang="en">
-
-	<head>
-		<meta charset="UTF-8" />
-		<title>SketchFlex</title>
-		<link href="https://fonts.googleapis.com/css?family=Raleway" rel="stylesheet" />
-		<link rel="stylesheet" href="css/style.css" />
-		<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-	</head>
-
-	<body>
-	  <div id = "container">
-	  		<div class="navbar">
-		    <div id="beginNav">
-			<div id="pageName">SketchFlex</div>
-			<a href="javascript:void(0);" id="hamburger"> <i
-				class="fa fa-bars"></i>
-			</a>
+<head>
+	<meta charset="UTF-8" />
+	<title>SketchFlex</title>
+	<link href="https://fonts.googleapis.com/css?family=Raleway" rel="stylesheet" />
+	<link rel="stylesheet" href="css/style.css" />
+	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+</head>
+<body>
+	<div id = "container">
+		<div class="navbar">
+			<div id="beginNav">
+				<div id="pageName">SketchFlex</div>
+				<a href="javascript:void(0);" id="hamburger"> 
+					<i class="fa fa-bars"></i>
+				</a>
 			</div>
-			<nav id="endNav">
-		
-			</nav>	
+			<nav id="endNav"></nav>	
 		</div>
 		<div id = "mainhome">
-		    <div id = "images">
-		      <img id = "imageone" src = "./images/storyboard-template.jpg"></img>
-		      <img id = "imagetwo" src = "./images/storyboard-template.jpg"></img>
-		      <img id = "imagethree" src = "./images/storyboard-template.jpg"></img>
-		    </div>
-		    <div id = "sidebar">
-			<h1 class="splash-tagline">
-				Your Sketch.
-				<br />
-				Your Story.
-				<br />
-				Your Way.
-			</h1>
-			<div id="works"><a href="#">How It Works!</a></div>
+			<div id = "images">
+				<img id = "imageone" src = "./images/storyboard-template.jpg"></img>
+				<img id = "imagetwo" src = "./images/storyboard-template.jpg"></img>
+				<img id = "imagethree" src = "./images/storyboard-template.jpg"></img>
+			</div>
+			<div id = "sidebar">
+				<h1 class="splash-tagline">
+					Your Sketch.
+					<br/>
+					Your Story.
+					<br />
+					Your Way.
+				</h1>
+				<div id="works">
+					<a href="#">How It Works!</a>
+				</div>
 			</div>
 		</div>
 		<footer>
-		  	<ul id="footerlist">
-						<li><a href="#">About</a></li>
-						<li><a href="#">Contact</a></li>
+			<ul id="footerlist">
+			<li><a href="#">About</a></li>
+			<li><a href="#">Contact</a></li>
 			</ul>
 		</footer>
 	</div>
 	<script type="text/javascript" th:src="@{/js/stickyNav.js}"></script>
-	</body>
-	</html>
+</body>
+</html>
