+++ src/main/resources/templates/single-sketchdeck-template.html
@@ -1,24 +1,21 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org">
-
-  <head>
-  <link rel="stylesheet" type="text/css" th:href="@{/css/style.css}">
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-  <link href="https://fonts.googleapis.com/css?family=Raleway" rel="stylesheet" />
-  <title th:text="${sketchdeck.name}"></title>
-  </head>
-  <body>
-    <div id = "container">
+<head>
+	<link rel="stylesheet" type="text/css" th:href="@{/css/style.css}">
+	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+	<link href="https://fonts.googleapis.com/css?family=Raleway" rel="stylesheet" />
+	<title th:text="${sketchdeck.name}"></title>
+</head>
+<body>
+	<div id = "container">
 		<div class="navbar">
-		    <div id="beginNav">
-			<div id="pageName">SketchFlex</div>
-			<a href="javascript:void(0);" id="hamburger"> <i
-				class="fa fa-bars"></i>
-			</a>
+			<div id="beginNav">
+				<div id="pageName">SketchFlex</div>
+				<a href="javascript:void(0);" id="hamburger">
+					<i class="fa fa-bars"></i>
+				</a>
 			</div>
-			<nav id="endNav">
-		
-			</nav>	
+			<nav id="endNav"></nav>	
 		</div>
 		<h1 class = "sketchDeck-name" th:text="${sketchdeck.name}"></h1>
 		<div class = "sketchDeck-description" th:text="${sketchdeck.description}"></div>
@@ -26,16 +23,15 @@
 			<a class="name-and-image" th:each="sketch:${sketches}" th:href="@{/sketch(id=${sketch.id})}">
 				<h2 class="sketch-name" th:text="${sketch.name}"></h2>
 				<img th:src="@{${sketch.imageLocation}}">
-				
 			</a>
 		</section>
 		<footer>
-		  	<ul id="footerlist">
-						<li><a href="#">About</a></li>
-						<li><a href="#">Contact</a></li>
+			<ul id="footerlist">
+				<li><a href="#">About</a></li>
+				<li><a href="#">Contact</a></li>
 			</ul>
 		</footer>
 	</div>
 	<script type="text/javascript" th:src="@{/js/stickyNav.js}"></script>
-  </body>
+</body>
 </html>
