+++ src/main/resources/templates/sketchView.html
@@ -1,37 +1,34 @@
 <!DOCTYPE html>
+<!-- Do we use this template??? -->
 <html xmlns:th="http://www.thymeleaf.org">
-
 <head>
 	<link rel="stylesheet" type="text/css" th:href="@{/css/style.css}">
 	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
 	<link href="https://fonts.googleapis.com/css?family=Raleway" rel="stylesheet" />
-		<title></title>
+	<title></title>
 </head>
 <body onload="init()">
-			<div class="navbar">
-		    <div id="beginNav">
+	<div class="navbar">
+		<div id="beginNav">
 			<div id="pageName">SketchFlex</div>
-			<a href="javascript:void(0);" id="hamburger"> <i
-				class="fa fa-bars"></i>
+			<a href="javascript:void(0);" id="hamburger">
+				<i class="fa fa-bars"></i>
 			</a>
-			</div>
-			<nav id="endNav">
-		
-			</nav>	
 		</div>
+		<nav id="endNav"></nav>	
+	</div>
 	<h1>XI. Canvas</h1>
-				<p></p>
-				<input type="submit" value="Clear Sketchpad" onclick="clearCanvas(canvas, ctx);"/>
+	<p></p>
+	<input type="submit" value="Clear Sketchpad" onclick="clearCanvas(canvas, ctx);"/>
 	<div id="canvasContainer">
-				<canvas id="sketchpad" ></canvas>
+		<canvas id="sketchpad" ></canvas>
 	</div>
 	<footer>
-	  	<ul id="footerlist">
-					<li><a href="#">About</a></li>
-					<li><a href="#">Contact</a></li>
+		<ul id="footerlist">
+			<li><a href="#">About</a></li>
+			<li><a href="#">Contact</a></li>
 		</ul>
 	</footer>
-  </div>
 </body>
 <script type="text/javascript" th:src="@{/js/stickyNav.js}"></script>
 </html>
